pyproject.toml
[tool.poetry]
name = "premai"
version = "0.3.42"
description = "A client library for accessing Prem APIs"
authors = []
readme = "README.md"
packages = [
    {include = "premai"},
]
include = ["CHANGELOG.md", "premai/py.typed"]


[tool.poetry.dependencies]
python = "^3.8"
httpx = ">=0.20.0,<0.27.0"
attrs = ">=21.3.0"
python-dateutil = "^2.8.0"
typing_extensions = ">=4.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["F", "I", "UP"]
line-length = 120
